Fixing the last issues with the i80 component
